Revert "Reverts "Standardize on Test* widgets in *_tester.dart files (#182395)" (#182406)"
This reverts commit ab86f66e946e58335418ef2865320be591b8ba45.
