+++ packages/flutter/test/widgets/basic_test.dart
@@ -17,8 +17,8 @@ import 'package:flutter/rendering.dart';
 import 'package:flutter/widgets.dart';
 import 'package:flutter_test/flutter_test.dart';
 
+import 'button_tester.dart';
 import 'semantics_tester.dart';
-import 'utils.dart';
 import 'widgets_app_tester.dart';
 
 void main() {
