+++ packages/flutter/test/widgets/focus_traversal_test.dart
@@ -10,6 +10,7 @@ import 'package:flutter/services.dart';
 import 'package:flutter/widgets.dart';
 import 'package:flutter_test/flutter_test.dart';
 
+import 'button_tester.dart';
 import 'semantics_tester.dart';
 import 'utils.dart';
 import 'widgets_app_tester.dart';
