+++ packages/flutter/test/widgets/semantics_test.dart
@@ -10,8 +10,8 @@ import 'package:flutter/rendering.dart';
 import 'package:flutter/widgets.dart';
 import 'package:flutter_test/flutter_test.dart';
 
+import 'button_tester.dart';
 import 'semantics_tester.dart';
-import 'utils.dart';
 
 void main() {
   setUp(() {
