+++ packages/flutter/test/widgets/slivers_test.dart
@@ -7,7 +7,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter/rendering.dart';
 import 'package:flutter_test/flutter_test.dart';
 
-import 'list_tile_test_utils.dart';
+import 'list_tile_tester.dart';
 import 'semantics_tester.dart';
 
 Future<void> test(WidgetTester tester, double offset, {double anchor = 0.0}) {
